2022-07-05-drf/aluguel_veiculos/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
importlib-metadata==4.12.0
Markdown==3.3.7
mysqlclient==2.1.1
pycodestyle==2.8.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
zipp==3.8.0
